requirements.txt
altgraph==0.17.4
attrs==25.1.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
customtkinter==5.2.2
darkdetect==0.8.0
docopt==0.6.2
et_xmlfile==2.0.0
h11==0.14.0
idna==3.10
numpy==2.2.3
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
pefile==2023.2.7
pillow==11.1.0
pycparser==2.22
pyinstaller==6.12.0
pyinstaller-hooks-contrib==2025.1
pypiwin32==223
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2025.1
pywin32==308
pywin32-ctypes==0.2.3
requests==2.32.3
selenium==4.29.0
setuptools==75.8.2
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.29.0
trio-websocket==0.12.2
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
websocket-client==1.8.0
wsproto==1.2.0
xlwings==0.33.9
yarg==0.1.10
